requirements.txt
pip==24.1.2
Babel==2.15.0
MarkupSafe==2.1.5
PyYAML==6.0.1
Send2Trash==1.8.3
absl-py==2.1.0
anaconda==0.0.1.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.2
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
notebook-shim==0.2.4
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pygame==2.6.1
pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyttsx3==2.98
pytz==2024.2
pywin32==306
pywinpty==2.0.13
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.14
ultralytics-thop==2.0.9
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
inflect==7.3.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
ordered-set==4.1.0
tomli==2.0.1
typeguard==4.3.0
wheel==0.43.0
zipp==3.19.2
